Test material for audio analysis
While audio analysis is working nicely across the board there are a
couple of assets which trigger a hang in extraction on OS X exclusively.

The hang will occur in both single-package and multi-package runs where
the offending package/asset is not the first to be processed.

The same packages/assets are being handled without problems on Linux
boxes. Most other packages/assets are being handled without problems on
OS X as well. Minus eg. pcm_98185286-866a-4fed-befb-279b5a9adce8_.mxf (2
channels, 24 bps, no channel config label, edit rate 60).

For audio analysis dcp_inspect sets up a ramdisk and extracts to FIFOs
which sox picks up to do its thing. It's been confirmed that taking the
ramdisk out of the equation does not solve the problem.

It's a bit baffling, that's for sure.